Changed userID into long

+++ OneBook-new/app/src/main/java/ca/ualberta/c301w19t14/onebook/Owner.java
@@ -2,13 +2,13 @@ package ca.ualberta.c301w19t14.onebook;
 
 public class Owner extends User {
     
-    Owner(String username, String password, String email, long phone, int userID) {
+    Owner(String username, String password, String email, long phone, long userID) {
 
         super(username, password, email, phone, userID);
 
     } 
 
-    public boolean lendBook(Book book, int userID) {
+    public boolean lendBook(Book book, long userID) {
         //code to be added
         return true;
     }
